Revert "6 7 8 9 10 & 스칼라 옮김 ㅋ (#34)"
This reverts commit eff8d07b016ce7095e782bccf348f7cea80875ff.
